requirements.txt
django==4.2.7
pyTelegramBotAPI==4.14.0
djangorestframework==3.14.0
openpyxl==3.1.2
telebot~=0.0.5
requests~=2.31.0